+++ testData/org/elixir_lang/code_insight/parameter_info/prefix_match_usage.ex
@@ -1,0 +1,5 @@
+defmodule ParameterInfo.PrefixMatchUsage do
+  alias ParameterInfo.PrefixMatchDeclaration
+
+  PrefixMatchDeclaration.reduce(<caret>)
+end